+++ src/Components/ClusterHeader/ClusterHeader.js
@@ -1,10 +1,11 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 
 import { Grid, GridItem } from '@patternfly/react-core/dist/js/layouts/Grid';
 import { Stack, StackItem } from '@patternfly/react-core/dist/js/layouts/Stack';
 import { Title } from '@patternfly/react-core/dist/js/components/Title';
+import { Dropdown, DropdownToggle, DropdownItem } from '@patternfly/react-core';
 import Skeleton from '@redhat-cloud-services/frontend-components/Skeleton';
 import DateFormat from '@redhat-cloud-services/frontend-components/DateFormat/DateFormat';
 
@@ -12,6 +13,8 @@ import messages from '../../Messages';
 import { OneLineLoader } from '../../Utilities/Loaders';
 
 export const ClusterHeader = ({ clusterId, clusterData }) => {
+  const location = window.location;
+  const [isOpen, setIsOpen] = useState(false);
   const intl = useIntl();
   // subscribe to the cluster data query
   const {
@@ -20,9 +23,23 @@ export const ClusterHeader = ({ clusterId, clusterData }) => {
     data: cluster,
   } = clusterData;
 
+  const redirectOCM = (clusterId) => {
+    location.assign(
+      location.origin +
+        (location.pathname.includes('beta') ? `/beta` : '') +
+        `/openshift/details/${clusterId}`
+    );
+  };
+
+  const dropDownItems = [
+    <DropdownItem key="link" onClick={() => redirectOCM(clusterId)}>
+      <snap>{intl.formatMessage(messages.clusterDetailsRedirect)}</snap>
+    </DropdownItem>,
+  ];
+
   return (
     <Grid id="cluster-header" md={12} hasGutter>
-      <GridItem>
+      <GridItem span={8}>
         {isUninitializedCluster || isFetchingCluster ? (
           <Skeleton size="sm" />
         ) : (
@@ -35,6 +52,23 @@ export const ClusterHeader = ({ clusterId, clusterData }) => {
             {clusterData?.data?.report?.meta.cluster_name || clusterId}
           </Title>
         )}
+      </GridItem>
+      <GridItem span={4} id="cluster-header-dropdown">
+        <Dropdown
+          position="right"
+          onSelect={() => setIsOpen(!isOpen)}
+          autoFocus={false}
+          isOpen={isOpen}
+          toggle={
+            <DropdownToggle
+              id="toggle-id-2"
+              onToggle={(isOpen) => setIsOpen(isOpen)}
+            >
+              {intl.formatMessage(messages.dropDownActionSingleCluster)}
+            </DropdownToggle>
+          }
+          dropdownItems={dropDownItems}
+        />
       </GridItem>
       <GridItem>
         <Stack>