+++ CHANGELOG.md
@@ -1,3 +1,57 @@
+## [1.3.1](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.3.0...v1.3.1) (2022-03-29)
+
+
+### Bug Fixes
+
+* **Cluster reports table:** fix inability to sort if param first passed ([ad7e58c](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/ad7e58c4e8b19fcb616b68a3933b47992d611b86))
+
+# [1.3.0](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.2.0...v1.3.0) (2022-03-29)
+
+
+### Features
+
+* **Clusters list table:** filters edition modifies search params ([#196](https://github.com/RedHatInsights/ocp-advisor-frontend/issues/196)) ([8536160](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/85361607bcb4cffa0c1c9b65de315f64c855fc92))
+
+# [1.2.0](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.1.0...v1.2.0) (2022-03-28)
+
+
+### Features
+
+* **Single cluster page:** View in clusters button ([#192](https://github.com/RedHatInsights/ocp-advisor-frontend/issues/192)) ([2b73580](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/2b735804bb44f91fdad36a937e02dc1d90c63ea1))
+
+# [1.1.0](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.0.16...v1.1.0) (2022-03-28)
+
+
+### Bug Fixes
+
+* **tables:** revert the time-based refresh feature ([5961ef7](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/5961ef7c2cade9593ec446ea09b0af0eeba595a5))
+
+
+### Features
+
+* **Recommendations list table:** filters edition modifies search params ([#195](https://github.com/RedHatInsights/ocp-advisor-frontend/issues/195)) ([364501d](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/364501dd33f4161c5e2281208130ae97d6c3a496))
+
+## [1.0.17](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.0.16...v1.0.17) (2022-03-28)
+
+
+### Bug Fixes
+
+* **tables:** revert the time-based refresh feature ([5961ef7](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/5961ef7c2cade9593ec446ea09b0af0eeba595a5))
+
+## [1.0.17](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.0.16...v1.0.17) (2022-03-25)
+
+
+### Bug Fixes
+
+* **tables:** revert the time-based refresh feature ([5961ef7](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/5961ef7c2cade9593ec446ea09b0af0eeba595a5))
+
+## [1.0.17](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.0.16...v1.0.17) (2022-03-24)
+
+
+### Bug Fixes
+
+* **tables:** revert the time-based refresh feature ([5961ef7](https://github.com/RedHatInsights/ocp-advisor-frontend/commit/5961ef7c2cade9593ec446ea09b0af0eeba595a5))
+
 ## [1.0.17](https://github.com/RedHatInsights/ocp-advisor-frontend/compare/v1.0.16...v1.0.17) (2022-03-24)